+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This projec
 # 2.4.42 (11/30/17)
 - IDME-4215: Fix Succesful User Wallet Sign In
 - IDME-4216: Fix Failed User Wallet Sign In test
+- IDME-4217: Fix User verifies a Military Group Affiliation test
 
 # 2.4.41 (11/16/17)
 - IDME-3763: Wallet Dashboard: Verify Sub-Nav Bar